Consolidate arpack and arnoldi tests
